refactor : 회원가입 및 로그인 훅 리팩토링, 불필요한 파일 삭제
